+++ networks.py
@@ -1,0 +1,52 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from tqdm import tqdm # Displays a progress bar
+
+import torch
+from torch import nn
+from torch import optim
+import torch.nn.functional as F
+from torchvision import datasets, transforms
+from torch.utils.data import Dataset, Subset, DataLoader, random_split, TensorDataset
+
+from dataset import EnableDataset
+
+import pickle
+
+from sklearn.model_selection import KFold, StratifiedKFold,ShuffleSplit ,train_test_split
+
+import copy
+import os
+import random
+
+
+
+
+class Network(nn.Module):
+    def __init__(self,NUMB_CLASS):
+        super().__init__()
+        self.sclayer1 = nn.Sequential(
+            nn.Conv2d(51, 128, kernel_size=5, stride=1, padding=2),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=2, stride=2))
+        self.sclayer2 = nn.Sequential(
+            nn.Conv2d(128, 256, kernel_size=5, stride=1, padding=2),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=2, stride=2),
+            )
+        self.drop_out = nn.Dropout()
+        self.fc1 = nn.Linear( 4096, 2000)
+        # self.fc1 = nn.Linear( 8192, 2000)
+        # self.fc1 = nn.Linear( 20480, 2000)
+        self.fc2 = nn.Linear(2000, NUMB_CLASS)
+
+
+    def forward(self,x):
+        x = self.sclayer1(x)
+        x = self.sclayer2(x)
+        x = x.reshape(x.size(0), -1)
+        x = self.drop_out(x)
+        x = self.fc1(x)
+        x = self.fc2(x)
+
+        return x
